production-ready mainnet path: reserve-backed land contract, eligibility gating, fail-closed preflight
- UtopiaLandMainnet: finite pre-reserved Stock Token rewards, no sale without full backing
- UtopiaEligibility: compliance-multisig-gated allowlist with expiry
- eligibility enforced on buy, transfer, and claim
- fail-closed DeployMainnet script and preflight-mainnet.sh
- dashboard confirms receipts, verifies ownership, shows wallet stock balances
- mainnet stays inert until deployment, provider, reserves, eligibility flow supplied

+++ contracts/src/UtopiaLandMainnet.sol
@@ -259,7 +259,7 @@ contract UtopiaLandMainnet is ERC721, Ownable2Step, ReentrancyGuard {
         from = super._update(to, tokenId, auth);
 
         uint256 word = tokenId >> 8;
-        uint256 mask = 1 << (tokenId & 255);
+        uint256 mask = uint256(1) << (tokenId & 255);
         if (from == address(0)) _ownershipBits[word] |= mask;
         else _ownerBits[from][word] &= ~mask;
         if (to == address(0)) _ownershipBits[word] &= ~mask;
